- moved tag names and advisories to `globals.js` - deleted  `e2e-test-constnats.js`
