--- e2e/e2e-tests-constants.js
@@ -1,1 +1,0 @@
-export const tooltipSourceSelector = '.overlay-indicator__tooltip__sources > .overlay-indicator__tooltip__source';
