+++ e2e/npm.spec.js
@@ -1,5 +1,5 @@
 import { test, Expect } from '../fixtures.js';
-import { packageReportTagName } from '../src/utils/tag-names.js';
+import { packageReportTagName } from '../src/globals.js';
 
 test.describe('npm', () => {
   test('Package Report is visible in npmjs.com', async ({ page }) => {
