+++ src/globals.js
@@ -8,3 +8,7 @@ export default advisories;
 
 export const advisoriesNames = Object.keys(advisories);
 export const advisoriesCount = advisoriesNames.length;
+
+//tag names and selectors
+export const packageReportTagName = 'overlay-package-report';
+export const tooltipSourceSelector = '.overlay-indicator__tooltip__sources > .overlay-indicator__tooltip__source';
