--- src/utils/tag-names.js
@@ -1,1 +1,0 @@
-export const packageReportTagName = 'overlay-package-report';
